added MyProfile

+++ yelp/src/components/Ad.js
@@ -1,57 +1,60 @@
 import React, {useEffect} from "react";
+import axios from 'axios'
 import ComplexGrid from './ComplexGrid';
 import { connect } from 'react-redux';
-
+import { favoriteSave } from '../actions/favoritesAction';
 import { fetchAd } from '../actions/adActions';
 
 
 
-// export default class Ad extends React.Component {
+export class Ad extends React.Component {
+ 
+  constructor(props) {
+    
+    super(props);
+    this.state = {
+      ad: null
+    };
+    console.log(`Ad`,this.props.favoriteSave)
+  }
+
+  // const Ad = (props) => {
+  //   console.log(`Ad props`, props)
+  //   const [ad, setAd] = ([])
   
-//   constructor(props) {
-//     console.log(`ad`, props)
-//     super(props);
-//     this.state = {
-//       ad: null
-//     };
-//   }
-
-  const Ad = (props) => {
-    console.log(`Ad props`, props)
-    const [ad, setAd] = ([])
   
-  
-  // //recieve from props id of ad that was clicked on, coming from ComplexGrid
-  // componentDidMount() {
-  //   console.log(this.props.match.params.id)
-  //   this.fetchMovie(this.props.match.params.id);
-  // }
+  //recieve from props id of ad that was clicked on, coming from ComplexGrid
+  componentDidMount() {
+    console.log(this.props.match)
+    this.fetchMovie(this.props.match.params.id);
+  }
 
-  // //if id is not = to 
-  // componentWillReceiveProps(newProps) {
-  //   if (this.props.match.params.id !== newProps.match.params.id) {
-  //     this.fetchMovie(newProps.match.params.id);
-  //   }
-  // }
+  //if id is not = to 
+  componentWillReceiveProps(newProps) {
+    if (this.props.match.params.id !== newProps.match.params.id) {
+      this.fetchMovie(newProps.match.params.id);
+    }
+  }
 
-  useEffect(() => {
-    props.fetchAd()
-  }, [])
+  // useEffect(() => {
+  //   props.fetchAd()
+  // }, [])
 
-      if (props.isFetching) {
-        return <h2>Loading Resturants Facts...</h2>;
-      }
+  //     if (props.isFetching) {
+  //       return <h2>Loading Resturants Facts...</h2>;
+  //     }
 
       
 
-  // fetchMovie = id => {
-  //   axios
-  //     .get(`https://cors-anywhere.herokuapp.com/https://cat-fact.herokuapp.com/facts/movies/${id}`)
-  //     .then(res => this.setState({ ad: res.data }))
-  //     .catch(err => console.log(err.response));
-  // };
+  fetchMovie = id => {
+    axios
+      .get(`https://yelp-feelers-be.herokuapp.com/reviews/${id}`)
+      .then(res => this.setState({ ad: res.data.review }))
+      // .then(res => console.log(res.data.review))
+      .catch(err => console.log(err.response));
+  };
 
-  const saveMovie = () => {
+  saveMovie = () => {
     const addToSavedList = this.props.addToSavedList;
     addToSavedList(this.state.ad);
   };
@@ -66,42 +69,49 @@ import { fetchAd } from '../actions/adActions';
   //     .catch(err => console.log(err.response));
   // };
 
-  // render() {
-  //   if (!this.state.ad) {
-  //     return <div>Loading movie information...</div>;
-  //   }
+  render() {
+    if (!this.state.ad) {
+      return <div>Loading movie information...</div>;
+    }
 
     return (
       <div className="save-wrapper">
-        <ComplexGrid movie={this.state.ad} />
-        <div className="save-button" onClick={this.saveMovie}>
-          Save
-        </div>
-        <button
+        <ComplexGrid fact={this.state.ad} />
+        <button className="save-button" onClick={this.props.favoriteSave}>
+          Save to Favorites
+        </button>
+        {/* <button
         onClick={() => this.props.history.push(`/update-movie/${this.state.ad.id}`)}
       >
         Edit Movie
       </button>
       <button onClick={this.deleteItem} className="md-button">
         Delete Item
-      </button>
+      </button> */}
       </div>
     );
   }
+}
 
-
-  const mapStateToProps = state => {
-    return {
-      ad: state.adReducer.ad,
-      isFetching: state.adReducer.isFetching,
-      error: state.adReducer.error
-    };
-  };
-  
-  export default connect(
-    mapStateToProps,
-    { fetchAd }
+export default connect(
+    null,
+    { favoriteSave }
   )(Ad);
+
+
+
+  // const mapStateToProps = state => {
+  //   return {
+  //     ad: state.adReducer.ad,
+  //     isFetching: state.adReducer.isFetching,
+  //     error: state.adReducer.error
+  //   };
+  // };
+  
+  // export default connect(
+  //   mapStateToProps,
+  //   { fetchAd }
+  // )(Ad);
 
 
 
@@ -111,30 +121,4 @@ import { fetchAd } from '../actions/adActions';
 
   
 
-  //if state is null, loading movie
-//   render() {
-   
-//     );
-//   }
-// }
 
-
-// if (!this.state.movie) {
-//   return <div>Loading movie information...</div>;
-// }
-
-// return (
-//   <div className="save-wrapper">
-//     <MovieCard movie={this.state.movie} />
-//     <div className="save-button" onClick={this.saveMovie}>
-//       Save
-//     </div>
-//     <button
-//     onClick={() => this.props.history.push(`/update-movie/${this.state.movie.id}`)}
-//   >
-//     Edit Movie
-//   </button>
-//   <button onClick={this.deleteItem} className="md-button">
-//     Delete Item
-//   </button>
-//   </div>
